feat: Car 테스트 코드 구현
- 자동차 생성 및 예외 테스트 코드 구현
