+++ src/test/kotlin/model/RacingGameTest.kt
@@ -3,7 +3,6 @@ package model
 import org.junit.jupiter.api.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import org.assertj.core.api.Assertions.assertThat
 
 class RacingGameTest {
     private lateinit var racingGame: RacingGame
@@ -17,30 +16,6 @@ class RacingGameTest {
     @ValueSource(strings = ["", " ", "car12345", "1", "car1, car1", "car1, car12345"])
     fun `자동차 생성 예외 테스트`(input: String) {
         assertThrows<IllegalArgumentException> { racingGame.generateCars(input) }
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = [4, 5, 6, 7, 8, 9])
-    fun `자동차 전진 테스트`(randomValue: Int) {
-        val car = Car("carA")
-        car.tryForward(randomValue)
-
-        val output = car.getStep()
-        val expected = "carA : -"
-
-        assertThat(output).isEqualTo(expected)
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = [0, 1, 2, 3])
-    fun `자동차 전진 실패 테스트`(randomValue: Int) {
-        val car = Car("carA")
-        car.tryForward(randomValue)
-
-        val output = car.getStep()
-        val expected = "carA : "
-
-        assertThat(output).isEqualTo(expected)
     }
 
     @ParameterizedTest
